FIX Ensure the identifier is set before trying to populate details from the API

+++ src/Model/Resource.php
@@ -47,7 +47,7 @@ class Resource extends DataObject
      */
     public function onAfterWrite()
     {
-        if ($this->isChanged('Identifier')) {
+        if ($this->Identifier && $this->isChanged('Identifier')) {
             $this->Fields()->each(function (ResourceField $field) {
                 $field->delete();
             });
@@ -74,7 +74,7 @@ class Resource extends DataObject
      */
     public function onBeforeWrite()
     {
-        if ($this->isChanged('Identifier')) {
+        if ($this->Identifier && $this->isChanged('Identifier')) {
             /** @var ResourcePopulatorInterface $populator */
             $populator = Injector::inst()->get(ResourcePopulatorInterface::class);
             $populator->populateMetadata($this);
